test: move tilegroup tests to hand module
These tests are actually testing the `Hand::new()` function and not the
TileGroup functions directly so it's appropriate to store these tests
under the `hand` module.
