+++ src/main.rs
@@ -302,163 +302,9 @@ fn main() {
 
 #[cfg(test)]
 mod test {
-    use mahc::fu::calculate_total_fu_value;
     use mahc::hand::error::HandErr;
-    use mahc::hand::Hand;
-    use mahc::suit::Suit;
-    use mahc::tile_group::GroupType;
 
     use super::*;
-
-    #[test]
-    fn identify_pair() {
-        let out = Hand::new(
-            vec![
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.pairs()[0].value, "S");
-        assert_eq!(out.pairs()[0].group_type, GroupType::Pair);
-        assert_eq!(out.pairs()[0].suit, Suit::Wind);
-        assert!(!out.pairs()[0].isopen);
-    }
-
-    #[test]
-    #[allow(non_snake_case)]
-    fn identify_tilegroup_closed_wind_trip_SSS() {
-        let out = Hand::new(
-            vec![
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.triplets()[0].value, "S");
-        assert_eq!(out.triplets()[0].group_type, GroupType::Triplet);
-        assert_eq!(out.triplets()[0].suit, Suit::Wind);
-        assert!(!out.triplets()[0].isopen);
-    }
-
-    #[test]
-    #[allow(non_snake_case)]
-    fn identify_tilegroup_open_wind_kan_EEEE() {
-        let out = Hand::new(
-            vec![
-                "EEEEwo".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.kans()[0].value, "E");
-        assert_eq!(out.kans()[0].group_type, GroupType::Kan);
-        assert_eq!(out.kans()[0].suit, Suit::Wind);
-        assert!(out.kans()[0].isopen);
-    }
-
-    #[test]
-    #[allow(non_snake_case)]
-    fn identify_tilegroup_closed_dragon_kan_RRRR() {
-        let out = Hand::new(
-            vec![
-                "rrrrd".to_string(),
-                "rrrrd".to_string(),
-                "SSw".to_string(),
-                "rrrrd".to_string(),
-                "rrrd".to_string(),
-            ],
-            "rd".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.kans()[0].value, "r");
-        assert_eq!(out.kans()[0].group_type, GroupType::Kan);
-        assert_eq!(out.kans()[0].suit, Suit::Dragon);
-        assert!(!out.kans()[0].isopen);
-    }
-
-    #[test]
-    fn identify_tilegroup_closed_manzu_trip_111() {
-        let out = Hand::new(
-            vec![
-                "111m".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.triplets()[0].value, "1");
-        assert_eq!(out.triplets()[0].group_type, GroupType::Triplet);
-        assert_eq!(out.triplets()[0].suit, Suit::Manzu);
-        assert!(!out.triplets()[0].isopen);
-    }
-
-    #[test]
-    fn identify_tilegroup_closed_souzu_seq_789() {
-        let out = Hand::new(
-            vec![
-                "789s".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.sequences()[0].value, "7");
-        assert_eq!(out.sequences()[0].group_type, GroupType::Sequence);
-        assert_eq!(out.sequences()[0].suit, Suit::Souzu);
-        assert!(!out.sequences()[0].isopen);
-    }
-
-    #[test]
-    fn identify_tilegroup_open_pinzu_234() {
-        let out = Hand::new(
-            vec![
-                "234po".to_string(),
-                "SSSw".to_string(),
-                "SSw".to_string(),
-                "SSSw".to_string(),
-                "SSSw".to_string(),
-            ],
-            "Sw".to_string(),
-            "3s".to_string(),
-            "3s".to_string(),
-        )
-        .unwrap();
-        assert_eq!(out.sequences()[0].value, "2");
-        assert_eq!(out.sequences()[0].group_type, GroupType::Sequence);
-        assert_eq!(out.sequences()[0].suit, Suit::Pinzu);
-        assert!(out.sequences()[0].isopen);
-    }
 
     #[test]
     fn no_han_for_calc() {
